+++ src/main/java/com/kookdonge/kookdonge_server/auth/service/UserService.java
@@ -101,16 +101,4 @@ public class UserService {
                 refreshToken);
     }
 
-    // AuthInterceptor에서 사용하기 위해서 만듬
-    public Long getUserIdByExternalUserId(String externalUserId){
-        UserEntity savedUserEntity = userRepository.findByExternalUserId(externalUserId)
-                .orElseThrow(() -> new CustomException(AuthExceptionCode.USER_NOT_FOUND));
-        return savedUserEntity.getUserId();
-    }
-    public Long getClubIdByUserId(Long userId){
-        UserEntity savedUserEntity = userRepository.findById(userId)
-                .orElseThrow(() -> new CustomException(AuthExceptionCode.USER_NOT_FOUND));
-        return savedUserEntity.getClubId();
-    }
-
 }
